feat: Update frontend appPort to 80 and add logLevel in frontend.bicep
The code changes in `frontend.bicep` modify the `appPort` property of the `frontend` resource from 3000 to 80. Additionally, a new property `logLevel` is added with the value 'info'. These changes aim to configure the frontend application to use port 80 and enable logging with the specified log level.

Recent user commits indicate a focus on modifying the infrastructure deployment, including adding an APIM module and renaming the `publisherEmail` parameter to `publisherEmailAddress` in `main.bicep`. However, these changes are not directly related to the modifications in `frontend.bicep`.

Signed-off-by: Will Velida <willvelida@hotmail.co.uk>