refactor: Update names and variables in serviceBus.bicep and cosmosDb.bicep
The code changes in `serviceBus.bicep` update the `topicName` variable from 'orders' to 'tasksavedtopic' to reflect the new functionality of the Service Bus. Additionally, in `cosmosDb.bicep`, the `databaseName` variable is changed from 'ServerlessStoreDB' to 'tasksmanagerdb' and the `containerName` variable is changed from 'orders' to 'taskscollection' to align with the updated naming conventions.